RAG/requirements.txt
faiss-cpu
python-docx
sentence-transformers
requests
beautifulsoup4
